sdk/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentops-monitor"
version = "0.1.2"
description = "Monitoring SDK for AI agents built with Google ADK"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "AgentOps Monitor Team", email = "contact@agentops-monitor.dev"}
]
maintainers = [
    {name = "AgentOps Monitor Team", email = "contact@agentops-monitor.dev"}
]
keywords = ["ai", "agents", "monitoring", "observability", "adk", "google-adk", "tracing"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "requests>=2.31.0",
    "google-adk>=1.13.0",
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
a2a = ["a2a-sdk>=0.3.16"]
dev = ["build>=1.0.0", "twine>=4.0.0", "pytest>=7.0.0"]

[project.urls]
Homepage = "https://github.com/sayandas24/agentops-monitor"
Repository = "https://github.com/sayandas24/agentops-monitor"
Documentation = "https://github.com/sayandas24/agentops-monitor#readme"
"Bug Tracker" = "https://github.com/sayandas24/agentops-monitor/issues"

[tool.setuptools]
packages = ["agentops_monitor", "agentops_monitor.adk"]
